tensorflow/requirements.txt
tensorflow
keras-cv
pillow
numpy
